refactor(provider-canalplus): rename d17 channel to cstar
The D17 channel rebranded to CStar in 2016 and now lives at
https://www.canalplus.com/chaines/cstar. Rename the embedded
Canal+ sub-provider accordingly so identifiers, grab-config entries
and documentation match the current channel name.

- plugins/canalPlus: rename D17Conf/D17PluginManager(+Test) to
  CStarConf/CStarPluginManager(+Test); provider id "d17" -> "cstar";
  HOME_URL points to the new canalplus.com home; remaining legacy
  endpoint strings get the d17 -> cstar substitution.
- application/consoleView/grabconfig.xml: rename channel id
  d17 -> cstar and category "Top D17" -> "Top CStar"
  (file kept as UTF-16 BE with BOM, untouched elsewhere).
- canalPlus offline fixture baseline + test: provider family entry
  d17 -> cstar.
- docs (README, provider-inventory, dev-tracker, dev-tracker.json):
  surface CStar with an "(ex D17)" historical label so legacy
  grab-config references stay discoverable.
