--- plugins/canalPlus/test/com/dabi/habitv/provider/canalplus/D17PluginManagerTest.java
@@ -1,14 +1,0 @@
-package com.dabi.habitv.provider.canalplus;
-
-import org.junit.Test;
-
-import com.dabi.habitv.plugintester.BasePluginProviderTester;
-
-public class D17PluginManagerTest extends BasePluginProviderTester {
-
-	@Test
-	public final void testProviderD17() throws InstantiationException, IllegalAccessException {
-		testPluginProvider(D17PluginManager.class, true);
-	}
-
-}
