+++ plugins/canalPlus/test/resources/fixtures/canalplus/fixture-baseline.txt
@@ -1,3 +1,3 @@
-providerFamily=canalPlus,d8,d17
+providerFamily=canalPlus,d8,cstar
 network=disabled
-note=Fixture placeholder for split Canal+/D8/D17 endpoint rewrite planning.
+note=Fixture placeholder for split Canal+/D8/CStar endpoint rewrite planning.
